Fixing some type issues: 	modified:   api.ipynb 	modified:   src/.ipynb_checkpoints/density_matrix-checkpoint.jl 	modified:   src/.ipynb_checkpoints/entanglement_functions-checkpoint.jl 	modified:   src/.ipynb_checkpoints/random_positive_matrix-checkpoint.jl 	modified:   src/.ipynb_checkpoints/unitary_transfs-checkpoint.jl 	modified:   src/density_matrix.jl 	modified:   src/entanglement_functions.jl 	modified:   src/random_positive_matrix.jl 	modified:   src/unitary_transfs.jl

+++ src/random_positive_matrix.jl
@@ -1,7 +1,7 @@
 using RandomMatrices
 using LinearAlgebra
 
-DimType=Int64
+DimType=Int
 
 function random_pure_rho(dimension::DimType)
     psi = randn(dimension) + randn(dimension)*1im
